+++ lawyerist-partner-dashboards.php
@@ -11,24 +11,33 @@ Author URI: https://lawyerist.com
 
 if ( !defined( 'ABSPATH' ) ) exit;
 
-define( 'PARTNER_DASHBOARDS_PLUGIN_VERSION', '0.1.0' );
+/**
+* Constants
+*/
+define( 'LPD_PLUGIN_VERSION', '0.1.0' );
+define( 'LPD_PAGE_SLUG', 'partner-dashboard' );
 // define( 'REST_AUTH_TOKEN', '' );
 
-require_once( plugin_dir_path( __FILE__ ) . 'common/partner-dashboards-common.php' );
-require_once( plugin_dir_path( __FILE__ ) . 'frontend/partner-dashboards-frontend.php' );
+require_once( plugin_dir_path( __FILE__ ) . 'common/lpd-common.php' );
 
 if ( is_admin() ) {
-	require_once( plugin_dir_path( __FILE__ ) . 'admin/partner-dashboards-admin.php' );
+	require_once( plugin_dir_path( __FILE__ ) . 'admin/lpd-admin.php' );
 }
 
-function frontend_stylesheet() {
-	wp_enqueue_style( 'partner-dashboards-frontend-css', plugins_url( 'frontend/partner-dashboards-frontend.css', __FILE__ ) );
-}
+if ( !is_admin() ) {
 
-add_action( 'wp_enqueue_scripts', 'frontend_stylesheet' );
+	require_once( plugin_dir_path( __FILE__ ) . 'frontend/lpd-frontend.php' );
+
+	function lpd_frontend_stylesheet() {
+		wp_enqueue_style( 'lpd-frontend-css', plugins_url( 'frontend/lpd-frontend.css', __FILE__ ) );
+	}
+
+	add_action( 'wp_enqueue_scripts', 'lpd_frontend_stylesheet' );
+
+}
 
 /**
 * Includes the activation functions to create a Partner post type and
 * Partner Dashboards page if one doesn't already exist.
 */
-include_once dirname( __FILE__ ) . '/activation.php';
+include_once dirname( __FILE__ ) . '/lpd-activation.php';